기초관리>매장정보관리>매장정보관리  - 매장환경복사 '상품' 복사시 선택한 매장의 정보를 복사함    └ 취급상품은 단독매장 제외

+++ src/main/java/kr/co/solbipos/store/manage/storemanage/service/impl/StoreManageServiceImpl.java
@@ -378,9 +378,13 @@ public class StoreManageServiceImpl implements StoreManageService{
                         procCnt += mapper.insertStoreSdselGroup(storeManageVO);
                         procCnt += mapper.insertStoreSdselClass(storeManageVO);
                         procCnt += mapper.insertStoreSdselProd(storeManageVO);
-                        // 취급상품
-                        procCnt += mapper.deleteStoreProductStore(storeManageVO);
-                        procCnt += mapper.insertStoreProductStore(storeManageVO);
+
+                        // 프랜차이즈매장
+                        if(!storeManageVO.getHqOfficeCd().equals("00000")) {
+                            // 취급상품
+                            procCnt += mapper.deleteStoreProductStore(storeManageVO);
+                            procCnt += mapper.insertStoreProductStore(storeManageVO);
+                        }
                     }
 
                     // 판매가 복사
